+++ netpyne/simFuncs.py
@@ -317,6 +317,13 @@ def setupRecording ():
     sim.pc.spike_record(-1, sim.simData['spkt'], sim.simData['spkid']) # -1 means to record from all cells on this node
 
     # stim spike recording
+    if isinstance(sim.cfg['analysis']['plotRaster'],dict) and 'include' in sim.cfg['analysis']['plotRaster']:
+        netStimPops = [pop.tags['popLabel'] for pop in sim.net.pops if pop.tags['cellModel']=='NetStim']+['allNetStims']
+        for item in sim.cfg['analysis']['plotRaster']['include']:
+            if item in netStimPops: 
+                sim.cfg['recordStim'] = True
+                break
+                  
     if sim.cfg['recordStim']:
         sim.simData['stims'] = {}
         for cell in sim.net.cells: