+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 - Added option to save figure, save figure data, specify figure size, and show figure to all analysis/plotting functions (issue #39)
 
+- Fixed bug when generating function-based random probability values -- made random stream independent 
+
 ## Version 0.4.8
 
 - Removed framework and init modules, and consolidated in sim module (so just need: 'from netpyne import sim')